Write operating BeesBeesBees source code
Things to remember from this sprint:
(1)use .call method within the scope of the
psuedoclassical constructor in order to have
access to the properties and methods of the
superclass.

(2)Set the prototype to point to an instance
of the superclass prototype. Avoid having the
prototype point to an instance of 'new
Superclass' since the new keyword requires
any arguments of the instance to have been
defined upon inception.

(3)Don't forget to set the prototype's
constructor to point back at the constructor
function.